backend/requirements.txt
fastapi
uvicorn
pydantic
requests
sqlalchemy
psycopg2-binary
python-dotenv
passlib[bcrypt]
roboflow
pillow
python-multipart
